+++ template/tests/test_MODULE__NAME.py
@@ -1,8 +1,8 @@
-"""Tests for PROJECT."""
+"""Tests for PROJECT__NAME."""
 
 from typer.testing import CliRunner
 
-from MODULE.cli import app
+from MODULE__NAME.cli import app
 
 runner = CliRunner()
 
@@ -10,4 +10,4 @@ runner = CliRunner()
 def test_hello() -> None:
     result = runner.invoke(app)
     assert result.exit_code == 0
-    assert "Hello from PROJECT!" in result.output
+    assert "Hello from PROJECT__NAME!" in result.output